+++ backend/src/test/java/hanglog/integration/service/CommunityServiceIntegrationTest.java
@@ -9,8 +9,8 @@ import hanglog.community.dto.response.CommunityTripListResponse;
 import hanglog.community.dto.response.CommunityTripResponse;
 import hanglog.community.service.CommunityService;
 import hanglog.expense.service.ExpenseService;
-import hanglog.trip.domain.repository.CustomDayLogRepositoryImpl;
-import hanglog.trip.domain.repository.CustomTripCityRepositoryImpl;
+import hanglog.trip.infrastructure.CustomDayLogRepositoryImpl;
+import hanglog.trip.infrastructure.CustomTripCityRepositoryImpl;
 import hanglog.trip.dto.request.PublishedStatusRequest;
 import hanglog.trip.dto.request.TripCreateRequest;
 import hanglog.trip.dto.response.TripDetailResponse;
