+++ backend/src/test/java/hanglog/integration/service/LikeServiceIntegrationTest.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import hanglog.community.domain.repository.LikeRepository;
 import hanglog.community.dto.request.LikeRequest;
 import hanglog.community.service.LikeService;
-import hanglog.trip.domain.repository.CustomDayLogRepositoryImpl;
-import hanglog.trip.domain.repository.CustomTripCityRepositoryImpl;
+import hanglog.trip.infrastructure.CustomDayLogRepositoryImpl;
+import hanglog.trip.infrastructure.CustomTripCityRepositoryImpl;
 import hanglog.trip.service.TripService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
